Reduce the cell size down to 8x8
Double the dimensions of the grid.

+++ cppconway/asm/fast_draw.h
@@ -1,1 +1,1 @@
-extern "C" void fast_draw_strip(uint16 *strip_src, uint16 *dest);
+extern "C" void fast_draw_strip(uint8 *strip_src, uint16 *dest);
